+++ Dado.hpp
@@ -7,21 +7,21 @@ class Dado {
     public:
         Dado(
             const std::string& nome,
-            const float idade,
+            const int idade,
             const float altura,
             const float peso,
             const char sexo
         );
 
         std::string getNome() const;
-        float getIdade() const;
+        int getIdade() const;
         float getAltura() const;
         float getPeso() const;
         char getSexo() const;
 
     private:
         std::string nome;
-        float idade;
+        int idade;
         float altura;
         float peso;
         char sexo;
