+++ ExtrairEstatisticasCSV.cpp
@@ -1,4 +1,5 @@
 #include "ExtrairEstatisticasCSV.hpp"
+#include "bibliotecas/csv.hpp"
 
 std::vector<Dado*> ExtrairEstatisticasCSV::converterParaClasseDado(std::ifstream& arq) const {
     std::vector<Dado*> dados;
@@ -9,7 +10,7 @@ std::vector<Dado*> ExtrairEstatisticasCSV::converterParaClasseDado(std::ifstream
     while (reader.read_row(row)) {
         dados.push_back(new Dado{
                 row["nome"].get<std::string>(),
-                row["idade"].get<float>(),
+                row["idade"].get<int>(),
                 row["altura"].get<float>(),
                 row["peso"].get<float>(),
                 row["sexo"].get<std::string>()[0]
